update deta storage, remove node-fetch shim
