+++ packages/denokv/deno.json
@@ -1,5 +1,5 @@
 {
-  "name": "@grammyjs/storages-denokv",
+  "name": "@grammyjs/storage-denokv",
   "version": "2.4.2",
   "exports": "./src/mod.ts",
   "tasks": {
