+++ packages/deta/deno.json
@@ -1,0 +1,9 @@
+{
+  "name": "@grammyjs/storage-deta",
+  "version": "2.4.2",
+  "exports": "./src/mod.ts",
+  "imports": {
+    "grammy": "https://lib.deno.dev/x/grammy@1.x/mod.ts"
+  },
+  "license": "./LICENSE"
+}
