+++ packages/deta/package.json
@@ -15,7 +15,7 @@
   "scripts": {
     "test:deno": "echo \"Error: no tests found\"",
     "prebuild": "rimraf dist",
-    "build": "deno2node tsconfig.cjs.json && deno2node tsconfig.esm.json && pnpm postbuild",
+    "build": "tsc -p tsconfig.cjs.json && tsc -p tsconfig.esm.json && pnpm postbuild",
     "postbuild": "tsx ../../tools/postBuildFixup.ts --path=dist"
   },
   "repository": {
@@ -34,12 +34,11 @@
     "url": "https://github.com/grammyjs/storages/issues"
   },
   "homepage": "https://github.com/grammyjs/storages/tree/main/packages/deta#readme",
-  "dependencies": {
-    "node-fetch": "^2.6.6"
-  },
   "devDependencies": {
-    "@types/node-fetch": "^2.6.11",
-    "grammy": "^1.21.1"
+    "grammy": "^1.30.0"
+  },
+  "peerDependencies": {
+    "grammy": "^1.30.0"
   },
   "keywords": [
     "grammy",
