--- packages/deta/src/deps.deno.ts
@@ -1,1 +1,0 @@
-export type { StorageAdapter } from 'https://lib.deno.dev/x/grammy@1.x/mod.ts';
