--- packages/deta/src/deps.node.ts
@@ -1,1 +1,0 @@
-export type { StorageAdapter } from 'grammy';
