--- packages/deta/src/shim.node.ts
@@ -1,1 +1,0 @@
-export { default as fetch, Headers } from 'node-fetch';
